requirements.txt
python-dotenv==1.0.0
supabase==2.3.5
resend==3.5.0
